Move CeloBackend from core to contracts
...and other import cleanups.

Goals:
* Prevent circular dependencies (being in core is really problematic in
  this regard)
* Avoid package aliases in imports
