+++ contracts/read_only_statedb.go
@@ -1,4 +1,4 @@
-package core
+package contracts
 
 import (
 	"math/big"
