+++ core/txpool/legacypool/celo_legacypool.go
@@ -4,7 +4,7 @@ import (
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core"
+	"github.com/ethereum/go-ethereum/contracts"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
 )
@@ -31,7 +31,7 @@ func (pool *LegacyPool) getBalances(address common.Address, currencies []common.
 }
 
 func (pool *LegacyPool) recreateCeloProperties() {
-	pool.celoBackend = &core.CeloBackend{
+	pool.celoBackend = &contracts.CeloBackend{
 		ChainConfig: pool.chainconfig,
 		State:       pool.currentState,
 	}
